parse_mysql/requirements.txt
boto3
botocore
jmespath
lxml
mysql-connector-python
python-dateutil
s3transfer
six
urllib3
